==> data_processing/2 - Generating MTurk Splits.ipynb <==
## 2) Generating MTurk Splits
Code to post the Amazon Mechanical Turk tasks programmatically. In Markdown mode to be extra sure this isn't re-run by accident.

import pandas as pd
df = pd.read_csv('../data/original_data_filtered.csv', index_col='index')
df['index'] = df.index

### Initial MTurk Test

# process for selecting initial MTurk data
test_for_turkers = df[df.img_found].groupby('confusion_label').apply(lambda x: x.sample(100))
chosen_records = test_for_turkers.index.levels[1]
test_for_turkers = df.loc[chosen_records]
test_for_turkers.to_csv('../data/initial_MTurk_test.csv', index_label='index')

### Second MTurk Task

initial_mturk_test = pd.read_csv('../data/initial_MTurk_test.csv', index_col='index')
print(f"Now dropping {len(initial_mturk_test)} rows from previous splits...")
df = df.drop(initial_mturk_test.index)

second_mturk_run = df[df.img_found].groupby('confusion_label').apply(lambda x: x.sample(3500))
chosen_records = second_mturk_run.index.levels[1]
second_mturk_run = df.loc[chosen_records]
second_mturk_run.to_csv('../data/second_MTurk_test.csv', index_label='index')
