==> data_processing/7 - Model Error Analysis.ipynb <==
## 7) Model Error Analysis
Visualize the errors of a trained model.

import numpy as np
import pandas as pd
import os, sys
from error_analysis import *

%load_ext autoreload
%autoreload 2

df_path = '../../bert_models/a_repeat/best_val_eval.csv'

df = pd.read_csv(df_path)

def convert(x):
    if pd.isnull(x): return None
    if type(x) is tuple: return x
    start_chunk, end_chunk = x.split(",")
    return (int(start_chunk[1:]), int(end_chunk[:-1]))

def extract_ans(tokenization, span):
    if span is None: return None
    return " ".join(tokenization[span[0]: span[1]+1])

if 'pred_span' in df.columns:
    df.pred_span = df.pred_span.apply(lambda x: convert(x))
    df.answer_intersection_span = df.answer_intersection_span.apply(lambda x: convert(x))
    df['predicted_answer'] = df.apply(lambda row: extract_ans(row['response_filtered'].split(), row['pred_span']), axis=1)

def display_func(row):
    print(f"{row.question} [{row.q_relevant}] {row.response_filtered} [{row.r_relevant}: {row.true_answer} | {row.predicted_answer}]")
   
analyze_df(df, df_path, display_func=display_func)
